Update html_actuator.js
Changing the particles so that they're actually mass-ordered, beginning at the electron and increasing up to the Higgs. Omitting the neutrinos, since (a) we don't know their mass ordering and (b) it fits nicely this way. Sadly, there is no room for the top ... but it's heavier than the Higgs anyway, so that seems fine to me.